% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init__apply_fixed_effects.R
\name{apply_fixed_effects}
\alias{apply_fixed_effects}
\title{Apply trial, donor, and weight FEs to popn.}
\usage{
apply_fixed_effects(popn, params)
}
\arguments{
\item{popn}{A population data.table}

\item{params}{A list of parameters}
}
\value{
A new population file with FEs applied to phenotypes
}
\description{
\code{apply_fixed_effects()} takes a population data.table and a
parameters list and applies all the fixed effects defined by \code{sim_X_fe}
according to the FE values contained in \code{FE_vals}.
}
