% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_links.R
\name{apply_links}
\alias{apply_links}
\title{Ensure linked traits are correctly applied}
\usage{
apply_links(params)
}
\arguments{
\item{params}{A list of parameters}
}
\value{
A new modified model parameters
}
\description{
Take `params` and ensure that linked traits are correct by
  setting `Sigma_G`, `priors`, and `fe_vals.`
}
