% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_km.R
\name{basic_km}
\alias{basic_km}
\title{KM plot}
\usage{
basic_km(popn, params)
}
\arguments{
\item{popn}{A population data.table}

\item{params}{A list of parameters}
}
\value{
A KM plot
}
\description{
Generate a basic KM plot from a single data.table
}
