% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting__check_L.R
\name{check_L}
\alias{check_L}
\title{Check Likelihoods}
\usage{
check_L(dataset = "fb-final", scen = 1)
}
\arguments{
\item{dataset}{A dataset string}

\item{scen}{A scenario number}
}
\value{
A plot of Likelihood values
}
\description{
Plot the likelihoods along MCMC chains
}
