% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_files.R
\name{cleanup_bici_files}
\alias{cleanup_bici_files}
\title{Clean up BICI files}
\usage{
cleanup_bici_files(params)
}
\arguments{
\item{params}{A list of parameters}
}
\description{
Remove old config files and results.
}
