% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodels.R
\docType{package}
\name{epimodels}
\alias{epimodels-package}
\alias{epimodels}
\title{epimodels: Epidemiological models}
\description{
Currently in development
}
\author{
\strong{Maintainer}: Jamie prentice \email{jamie.c.prentice@gmail.com}

Authors:
\itemize{
  \item Jamie prentice \email{jamie.c.prentice@gmail.com}
}

Other contributors:
\itemize{
  \item University of Edinburgh [copyright holder, funder]
}

}
