% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation__fix_fb_data.R
\name{fix_fb_data}
\alias{fix_fb_data}
\title{Fix FB Data}
\usage{
fix_fb_data(popn, params)
}
\arguments{
\item{popn}{A Population in data.table format}

\item{params}{A list of parameters}
}
\value{
An updated \code{popn}
}
\description{
Potentially reclassify donors as recipients
}
