% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_bici_states.R
\name{flatten_bici_states}
\alias{flatten_bici_states}
\title{Flatten BICI states}
\usage{
flatten_bici_states(dataset = "fb-test", name = "scen-1-1", bici_cmd = "inf")
}
\arguments{
\item{dataset}{Name of the datasets}

\item{name}{Name of the run (e.g. "scen-1-1")}

\item{bici_cmd}{bici_cmd: where do the output of BICI's simulations come from?
\itemize{
\item "inf": look for all 'state_\if{html}{\out{<n>}}.txt' files, and save to 'etc.rds'
\item "sim": look for all 'state_\if{html}{\out{<n>}}.txt' files, and save to 'etc-sim.rds'
\item "post-sim": look for a single 'state.txt' file and save to 'etc-ps.rds'
}}
}
\value{
An R list containing all BICI information
}
\description{
Take the BICI state files and create an \code{etc.rds} file that contains all the
parameters and traits in a data.table for easy access.
}
