% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init__generate_bici_script.R
\name{generate_bici_script}
\alias{generate_bici_script}
\title{Generate BICI script}
\usage{
generate_bici_script(popn, params)
}
\arguments{
\item{popn}{A population with traits and epidemic times}

\item{params}{A list of parameters}
}
\value{
A character vector of lines of BICI script
}
\description{
Generate BICI script
}
