% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_R0.R
\name{get_R0}
\alias{get_R0}
\title{Get R0}
\usage{
get_R0(popn)
}
\arguments{
\item{popn}{A Population with infection generations.}
}
\value{
An estimated R0
}
\description{
Calculate R0 from the output of a simulated epidemic using the ratio of Gen 1
to Gen 2 infectives.
}
