% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation__get_tmax.R
\name{get_tmax}
\alias{get_tmax}
\title{Get T_max}
\usage{
get_tmax(popn, params)
}
\arguments{
\item{popn}{A Population with event times}

\item{params}{A list of parameters}
}
\value{
The maximum model time
}
\description{
Find \code{tmax} for each trial, after possibly censoring the last proportion of
the population, given by \code{censor}, and rounded up to the next \code{time_step}
value.
}
