% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grm.R
\name{make_grm}
\alias{make_grm}
\title{Title}
\usage{
make_grm(popn, use_grm)
}
\arguments{
\item{popn}{A pedigree in data.table format}

\item{use_grm}{A string listing what type of A matrix to produce.
- "A" is a standard A matrix
- "A_nz" is A in sparse format
- "A_inv" is inverse A
- "A_inv_nz" is inverse A in sparse format}
}
\value{
A numerator relationship "A" Matrix
}
\description{
Title
}
