% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time_series.R
\name{make_time_series}
\alias{make_time_series}
\title{Make Time Series}
\usage{
make_time_series(popn, params)
}
\arguments{
\item{popn}{A population with appropriate epidemic event times}

\item{params}{A list of parameters}
}
\value{
A data.table time series
}
\description{
Make Time Series
}
