% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__model_SEIDR.R
\name{model_SEIDR}
\alias{model_SEIDR}
\title{Simulate an SEIDR model}
\usage{
model_SEIDR(popn, params)
}
\arguments{
\item{popn}{A population with trait values in data.table format}

\item{params}{A list of parameters}
}
\value{
A new popn with epidemic event times, generation, and infectors
}
\description{
Simulate an SEIDR model
}
