% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__model_SEIR.R
\name{model_SEIR}
\alias{model_SEIR}
\title{Simulate an SEIR model}
\usage{
model_SEIR(popn, params)
}
\arguments{
\item{popn}{A population with trait values in data.table format}

\item{params}{A list of parameters}
}
\value{
A new popn with epidemic event times, generation, and infectors
}
\description{
Simulate an SEIR model
}
