% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__model_SIDR.R
\name{model_SIDR}
\alias{model_SIDR}
\title{Simulate an SIDR model}
\usage{
model_SIDR(popn, params)
}
\arguments{
\item{popn}{A population with trait values in data.table format}

\item{params}{A list of parameters}
}
\value{
A new popn with epidemic event times, generation, and infectors
}
\description{
Simulate an SIDR model
}
