% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__model_SIR.R
\name{model_SIR}
\alias{model_SIR}
\title{Simulate an SIR model}
\usage{
model_SIR(popn, params)
}
\arguments{
\item{popn}{A population with trait values in data.table format}

\item{params}{A list of parameters}
}
\value{
A new popn with epidemic event times, generation, and infectors
}
\description{
Simulate an SIR model
}
