% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models__model_SIS.R
\name{model_SIS}
\alias{model_SIS}
\title{Simulate an SIS model}
\usage{
model_SIS(popn, params)
}
\arguments{
\item{popn}{A population with trait values in data.table format}

\item{params}{A list of parameters}
}
\value{
A new popn with epidemic event times, generation, and infectors
}
\description{
Simulate an SIS model
}
