% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation__patch_params.R
\name{patch_params}
\alias{patch_params}
\title{Patch Parameters}
\usage{
patch_params(params, trace_row = 0)
}
\arguments{
\item{params}{A list of parameters}

\item{trace_row}{An optional row number from a trace file}
}
\value{
An updated parameters list
}
\description{
This will read an existing fitted dataset and apply either the mean or median
parameter values, or sample from the posteriors. It saves the source so that
the corresponding BVs can be applied too (but we do that later).
}
