% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_ts.R
\name{plot_model}
\alias{plot_model}
\title{Plot Model time series}
\usage{
plot_model(popn, params)
}
\arguments{
\item{popn}{A population with epidemic event times}

\item{params}{A list of parameters}
}
\value{
A plot of the epidemic
}
\description{
Plot Model time series
}
