% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting__posteriors.R
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plot posterior distribution}
\usage{
plot_posteriors(
  dataset = "fb-final",
  scen = 1,
  rep = 1,
  ci = "hpdi",
  draw = "density"
)
}
\arguments{
\item{dataset}{name of the data set to be worked on (e.g. \code{"fb-final"})}

\item{scen}{scenario number}

\item{ci}{credible interval: \code{"hpdi"} is the highest posterior density
interval. Anything else gives the 95\% credible interval}

\item{draw}{choose between a \code{"density"} plot or a \code{"histogram"}}
}
\value{
a list containing various plots and subplots
}
\description{
Plot a grid of posterior distributions for parameters from a data set.
}
\examples{
`plot_posteriors("fb-final", 1)`
`plot_posteriors("fb-final", 1, "hpdi", "histogram")`
}
