% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init__remove_bici_fes.R
\name{remove_bici_fes}
\alias{remove_bici_fes}
\title{Remove BICI Fixed Effects}
\usage{
remove_bici_fes(params)
}
\arguments{
\item{params}{A list of parameters}
}
\value{
An updated \code{params} list
}
\description{
BICI doesn't use Trial or Donor FEs, so these need to be carefully adjusted.
}
