% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_pars.R
\name{rename_bici_pars}
\alias{rename_bici_pars}
\title{Rename BICI parameters}
\usage{
rename_bici_pars(pars)
}
\arguments{
\item{pars}{A character vector of BICI parameter names}
}
\value{
A new character vector of modified parameter names
}
\description{
Take BICI parameters and make them consistent with internal parameter names,
e.g.
\itemize{
\item "G_1" -> "Group effect 1"
\item "\\mu^weight1s" -> "weight1_s"
\item "\\Omega^gen_sg,sg" -> "cov_G_ss"
\item "\\omega^gen_sg,ig" -> "r_G_si"
}
}
