% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_pars.R
\name{rename_pars}
\alias{rename_pars}
\title{Rename parameters}
\usage{
rename_pars(pars)
}
\arguments{
\item{pars}{A character vector of parameter names}
}
\value{
A new character vector of modified parameter names
}
\description{
Turns the short name parameters into something suitable for printing, e.g.
\itemize{
\item "cov_G_ss" -> "Var G (sus)"
\item "LP" -> "Latent Period (days)"
\item "trial_i" -> "Trial Infectivity"
}
}
