% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation__set_groups.R
\name{set_groups}
\alias{set_groups}
\title{Set Groups}
\usage{
set_groups(popn, params)
}
\arguments{
\item{popn}{A pedigree data.table}

\item{params}{A list of parameters}
}
\value{
A new \code{popn} data.table.
}
\description{
Copy the \code{popn} DT and assign groups, donors and recipients, group effect,
and trials.
}
