% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init__set_phenotypes.R
\name{set_phenotypes}
\alias{set_phenotypes}
\title{Apply trial, donor, and weight FEs to popn.}
\usage{
set_phenotypes(popn, params)
}
\arguments{
\item{popn}{A population data.table}

\item{params}{A list of parameters}
}
\value{
A new population file with FEs applied to phenotypes
}
\description{
\code{set_phenotypes()} takes a population data.table and a
parameters list and applies all the fixed effects defined by \code{sim_X_fe}
according to the FE values contained in \code{FE_vals}.
}
