% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init__set_traits.R
\name{set_traits}
\alias{set_traits}
\title{Set call appropriate function depending on the \code{trait_source}}
\usage{
set_traits(popn, params)
}
\arguments{
\item{popn}{A pedigree data.table}

\item{params}{A list of parameters}
}
\value{
A new \code{popn} data.table.
}
\description{
Set call appropriate function depending on the \code{trait_source}
}
