% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation__simulate_epidemic.R
\name{simulate_epidemic}
\alias{simulate_epidemic}
\title{Simulate an Epidemic Model}
\usage{
simulate_epidemic(popn, params)
}
\arguments{
\item{popn}{A population with trait values in data.table format}

\item{params}{A list of parameters}
}
\value{
A new popn with epidemic event times, generation, and infectors
}
\description{
Calls the correct model based on \code{params$model_type}.
}
