% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up_periods.R
\name{tidy_up_periods}
\alias{tidy_up_periods}
\title{Tidy up periods}
\usage{
tidy_up_periods(params, protocol = NULL)
}
\arguments{
\item{params}{A list of parameters}

\item{protocol}{A list to override params}
}
\value{
A new `params` list
}
\description{
If modifying the `params` list (likely with a `protocol` list) touches any of
the TPs or shape parameters, make sure that these changes are reflected in
the rate parameters and the priors.
}
