% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen_priors.R
\name{widen_priors}
\alias{widen_priors}
\title{Widen Priors}
\usage{
widen_priors(priors)
}
\arguments{
\item{priors}{A data.table of priors}
}
\value{
A modified data.table of priors
}
\description{
Widen Priors
}
