% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init__apply_links.R
\name{apply_links}
\alias{apply_links}
\title{Ensure linked traits are correctly applied}
\usage{
apply_links(params)
}
\arguments{
\item{params}{A list of parameters}
}
\value{
A new modified model parameters
}
\description{
Take \code{params} and ensure that linked traits are correct by
setting \code{Sigma_G}, \code{priors}, and \code{fe_vals.}
}
