% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmax.R
\name{get_tmax}
\alias{get_tmax}
\title{Get T_max}
\usage{
get_tmax(popn, params)
}
\arguments{
\item{popn}{A Population with event times}

\item{params}{A list of parameters}
}
\value{
The maximum model time
}
\description{
Find `tmax` for each trial, after possibly censoring the last proportion of
the population, given by `censor`, and rounded up to the next `time_step`
value.
}
