% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrices.R
\name{make_matrices}
\alias{make_matrices}
\title{Construct cov_mat and Sigma matrices using var and cor components}
\usage{
make_matrices(
  model_traits = c("sus", "inf", "lat", "det", "tol"),
  use_traits = "sildt",
  vars = 1,
  cors = 0.2
)
}
\arguments{
\item{model_traits}{A character vector of traits, e.g. `c("sus", "inf", "tol)`}

\item{use_traits}{A string with the first letter of each trait being used, e.g. `sit`}

\item{vars}{Variances. Either
- A single value, e.g. `1`
- A named list, e.g. `list(sus = 1, inf = 1.5, default = 0)`}

\item{cors}{Correlations. Either
- A single value in (-1,1), e.g. `0.2`,
- a named list e.g. `list(si = 0.3, st = -0.2, default = 0)`}
}
\value{
A list containing Sigma matrices (for BICI) and covariance matrices (for R)
}
\description{
Construct cov_mat and Sigma matrices using var and cor components
}
\examples{
out <- make_matrices(model_traits, "sit",
                     vars = list(sus = 1, inf = 1.5, default = 0),
                     cors = list(si = 0.3, st = 0.2, default = 0))
}
