% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{make_parameters}
\alias{make_parameters}
\title{Generate a list of parameters}
\usage{
make_parameters(
  model_type = "SEIR",
  name = "scen-1-1",
  dataset = "testing",
  scenario = 1L,
  replicate = 1L,
  setup = "fb_12_rpw",
  group_layout = "fishboost",
  IEs = "sit",
  vars = 1,
  cors = 0.2,
  FEs = list(sex = "sit", age = "sit", weight = "sit"),
  group_effect = -1,
  sim_new_data = "r",
  protocol = NULL
)
}
\arguments{
\item{model_type}{Epidemic model: "SIR", "SIDR", "SEIR", "SEIDR"}

\item{name}{Name of scenario within dataset (default "scen-1-1")}

\item{dataset}{Name for overall dataset (default "testing")}

\item{scenario}{Integer for scenario}

\item{replicate}{Integer for replicate}

\item{setup}{Population layout}

\item{group_layout}{How to arrange individuals into groups: "random",
"striped", "family", "fishboost"}

\item{IEs}{Which traits are affected by individual genetic effects, either in
simulation or inference. This should correspond to the names of the model
traits used. Use a trait's first letter if used, or _ if unused. For linked
traits use the letter of the trait it's linked to. "", "all", or "none" are
also accepted, with the obvious meanings/}

\item{vars}{Variances: diagonal of covariance matrix, either
\itemize{
\item a single value, e.g. 1.0
\item a list wrapped vector of compatible length, e.g.
\itemize{
\item list(1, 1.5, 0, 0, 0.5)
}
\item a list wrapped named vector e.g.
\itemize{
\item list(sus = 1, inf = 1.5, tol = 0.5, default = 0)
}
}}

\item{cors}{Correlations for Sigma matrix, either
\itemize{
\item a single numerical value, e.g. 0.2
\item a named list-wrapped vector, e.g.
\itemize{
\item list(si = 0.3, st = -0.1, it = 0.2, default = 0)
}
\item a named list of compatible with the lower triangle order, e.g.
\itemize{
\item list(si, sl, sd, st, il, id, it, ld, lt, dt)
}
}}

\item{FEs}{A list of Fixed Effect names. Values should correspond to the full
list of model traits they apply to, e.g. for an SEIR model with traits
"silt" use:
list(sex = "silt",   # all traits used
trial = "ilt",  # susceptibility not used
weight = "sit") # latency not used}

\item{group_effect}{Include group effect if >= 0, simulate if appropriate and
check, ignore if < 0}

\item{sim_new_data}{Simulate new data in "r" or "bici". Alternatively, "no"
for using FB data, or "summary_sim" or "summary_inf" for getting data from
an existing dataset (specified in patches)}
}
\value{
A list containing all the parameters
}
\description{
Generate a complete list of parameters, factoring in arguments
provided. Values can be modified later if preferred, but be careful about
maintaining consistency.
}
