% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{make_parameters}
\alias{make_parameters}
\title{Generate a list of parameters}
\usage{
make_parameters(
  model_type = "SEIDR",
  name = "scen-1-1",
  dataset = "testing",
  scenario = 1L,
  replicate = 1L,
  setup = "fb_12_rpw",
  use_traits = "sit",
  vars = 1,
  cors = 0.2,
  group_layout = "fishboost",
  group_effect = -1,
  trial_fe = "none",
  donor_fe = "none",
  txd_fe = "none",
  weight_fe = "none",
  weight_is_nested = TRUE,
  sim_new_data = "bici"
)
}
\arguments{
\item{model_type}{Epidemic model: "SIR", "SIDR", "SEIR", "SEIDR"}

\item{name}{Name of scenario within dataset (default "scen-1-1")}

\item{dataset}{Name for overall dataset (default "testing")}

\item{scenario}{Integer for scenario}

\item{replicate}{Integer for replicate}

\item{setup}{Population layout}

\item{use_traits}{Which traits to use (this is going to be "clever") use
\itemize{
\item "all" or "none", or a subset using the first letter of each trait
\item "sildt" = "sus", "inf", "lat", "det", "tol"
}}

\item{vars}{Variances: diagonal of covariance matrix, either
\itemize{
\item a single value, e.g. 1.0
\item a list wrapped vector of compatible length, e.g.
\itemize{
\item list(1, 1.5, 0, 0, 0.5)
}
\item a list wrapped named vector e.g.
\itemize{
\item list(sus = 1, inf = 1.5, tol = 0.5, default = 0)
}
}}

\item{cors}{Correlations for Sigma matrix, either
\itemize{
\item a single numerical value, e.g. 0.2
\item a named list-wrapped vector, e.g.
\itemize{
\item list(si = 0.3, st = -0.1, it = 0.2, default = 0)
}
\item a named list of compatible with the lower triangle order, e.g.
\itemize{
\item list(si, sl, sd, st, il, id, it, ld, lt, dt)
}
}}

\item{group_layout}{How to arrange individuals into groups: "random",
"striped", "family", "fishboost"}

\item{group_effect}{Include group effect if >= 0, simulate if appropriate and
check, ignore if < 0}

\item{trial_fe}{List of fixed effects to check for Trial Should also simulate
these if not using FB dataset. A string with the first letters of traits,
e.g. "lidt"}

\item{donor_fe}{List of fixed effects to check for Donor}

\item{txd_fe}{List of fixed effects to check for Trial x Donor interaction}

\item{weight_fe}{List of fixed effects to check for Weight}

\item{weight_is_nested}{should weight be nested across trials? (default
TRUE),}

\item{sim_new_data}{Simulate new data in "r" or "bici". Alternatively, "no"
for using FB data, or "etc_sim" or "etc_inf" for getting data from an
existing dataset (specified in patches)}
}
\value{
A list containing all the parameters
}
\description{
Generate a complete list of parameters, factoring in arguments
provided. Values can be modified later if preferred, but be careful about
maintaining consistency.
}
