% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pedigree.R
\name{make_pedigree}
\alias{make_pedigree}
\title{Make Pedigree}
\usage{
make_pedigree(params)
}
\arguments{
\item{params}{A parameter list}
}
\value{
A pedigree data.table.
}
\description{
Use the `params` list to generate a new pedigree. If `setup` includes `"fb"`,
then instead load the relevant FB dataset.
}
